multiple calls handles, strted prefs

+++ FeelsLikeWeather/app/src/main/java/com/example/kylie/feelslikeweather/models/darkskypojos/Currently.java
@@ -28,7 +28,7 @@ public class Currently implements Serializable {
     private Double precipIntensityError;
     @SerializedName("precipProbability")
     @Expose
-    private Integer precipProbability;
+    private Double precipProbability;
     @SerializedName("precipType")
     @Expose
     private String precipType;
@@ -111,11 +111,11 @@ public class Currently implements Serializable {
         this.precipIntensityError = precipIntensityError;
     }
 
-    public Integer getPrecipProbability() {
+    public Double getPrecipProbability() {
         return precipProbability;
     }
 
-    public void setPrecipProbability(Integer precipProbability) {
+    public void setPrecipProbability(Double precipProbability) {
         this.precipProbability = precipProbability;
     }
